+++ app/src/main/java/com/fibelatti/photowidget/ui/ColoredShape.kt
@@ -8,29 +8,23 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.asComposePath
-import androidx.graphics.shapes.RoundedPolygon
-import androidx.graphics.shapes.toPath
 import com.fibelatti.photowidget.model.PhotoWidgetShapeBuilder
 
 @Composable
 fun ColoredShape(
-    polygon: RoundedPolygon,
+    shapeId: String,
     color: Color,
     modifier: Modifier = Modifier,
     content: @Composable BoxScope.() -> Unit = {},
 ) {
     Box(
         modifier = modifier.drawWithContent {
-            val sizedPolygon = PhotoWidgetShapeBuilder.resizeShape(
-                roundedPolygon = polygon,
-                width = size.width,
-                height = size.height,
-            )
-
             drawPath(
-                path = sizedPolygon
-                    .toPath()
-                    .asComposePath(),
+                path = PhotoWidgetShapeBuilder.getShapePath(
+                    shapeId = shapeId,
+                    width = size.width,
+                    height = size.height,
+                ).asComposePath(),
                 color = color,
             )