+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* New shape: star
+
 ### Changed
 
+* Gave the heart shape a bit of love
 * Enable border customization when using "fill widget area"
 
 ### Fixed